requirements.txt
requests 
beautifulsoup4
groq
SimplerLLM
selenium
webdriver-manager
google-auth 
google-auth-oauthlib 
google-auth-httplib2 
google-api-python-client
geocoder
opencage